requirements.txt
speedtest-cli>=2.1.3
geopy>=2.4.0
pandas>=1.3.0
pytest>=7.0.0
pytest-cov>=4.0.0
